Removed usage of data.table package
At some point following the publication of this study, the R package "data.table" was updated and that update lead to errors in the "lcfc_experiment_analysis.R" script. "data.table" was only used to read in the .csv files containing the raw data, so an alternative that does not use the package was implemented. No results have changed.